// src/app/model/pokemon.models.ts
// Generated by https://quicktype.io

export interface Pokemon {
  pokedexId:   number;
  generation:  number;
  category:    string;
  name:        NameClass;
  sprites:     Sprites;
  types:       Type[] | null;
  talents:     Talent[] | null;
  stats:       Stats | null;
  resistances: Resistance[] | null;
  evolution:   Evolution | null;
  height:      null | string;
  weight:      null | string;
  egg_groups:  EggGroup[] | null;
  sexe:        Sexe | null;
  catch_rate:  number | null;
  level_100:   number | null;
  forme:       Forme[] | null;
  next?:       null;
}

export enum EggGroup {
  Amorphe = "Amorphe",
  Aquatique1 = "Aquatique 1",
  Aquatique2 = "Aquatique 2",
  Aquatique3 = "Aquatique 3",
  Aérien = "Aérien",
  Draconique = "Draconique",
  Féerique = "Féerique",
  Humanoïde = "Humanoïde",
  Insectoïde = "Insectoïde",
  Minéral = "Minéral",
  Monstrueux = "Monstrueux",
  Métamorph = "Métamorph",
  Terrestre = "Terrestre",
  Végétal = "Végétal",
}

export interface Evolution {
  pre:   Next[] | null;
  next?: Next[] | null;
  mega:  Mega[] | null;
}

export interface Mega {
  orbe?:   string;
  sprites: Gmax;
  obre?:   string;
}

export interface Gmax {
  regular: string;
  shiny:   null | string;
}

export interface Next {
  pokedexId:  number;
  name:       string;
  condition?: string;
}

export interface Forme {
  alola?:  string;
  galar?:  string;
  hisui?:  string;
  paldea?: string;
}

export interface NameClass {
  fr: string;
  en: string;
  jp: string;
}

export interface Resistance {
  name:       NameEnum;
  multiplier: number;
}

export enum NameEnum {
  Acier = "Acier",
  Combat = "Combat",
  Dragon = "Dragon",
  Eau = "Eau",
  Feu = "Feu",
  Fée = "Fée",
  Glace = "Glace",
  Insecte = "Insecte",
  Normal = "Normal",
  Plante = "Plante",
  Poison = "Poison",
  Psy = "Psy",
  Roche = "Roche",
  Sol = "Sol",
  Spectre = "Spectre",
  Ténèbres = "Ténèbres",
  Vol = "Vol",
  Électrik = "Électrik",
}

export interface Sexe {
  male:   number;
  female: number;
}

export interface Sprites {
  regular: string;
  shiny:   null | string;
  gmax:    Gmax | null;
}

export interface Stats {
  hp:      number;
  atk:     number;
  def:     number;
  spe_atk: number;
  spe_def: number;
  vit:     number;
}

export interface Talent {
  name: string;
  tc:   boolean;
}

export interface Type {
  name:  NameEnum;
  image: string;
}